move parties endpoints to separate module
