+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/endpoints/ContractList.scala
@@ -1,3 +1,6 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 package com.daml.http
 package endpoints
