+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/endpoints/Parties.scala
@@ -1,0 +1,87 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.http
+package endpoints
+
+import akka.NotUsed
+import akka.http.scaladsl.model._
+import headers.`Content-Type`
+import akka.http.scaladsl.server
+import akka.http.scaladsl.server.Directives.extractClientIP
+import akka.http.scaladsl.server.{Directive, Directive0, PathMatcher, Route}
+import akka.http.scaladsl.server.RouteResult._
+import akka.stream.Materializer
+import akka.stream.scaladsl.{Flow, Source}
+import akka.util.ByteString
+import com.codahale.metrics.Timer
+import ContractsService.SearchResult
+import EndpointsCompanion._
+import Endpoints.ET
+import json._
+import util.Collections.toNonEmptySet
+import util.FutureUtil.either
+import util.Logging.{InstanceUUID, RequestID, extendWithRequestIdLogCtx}
+import com.daml.logging.LoggingContextOf.withEnrichedLoggingContext
+import scalaz.std.scalaFuture._
+import scalaz.syntax.std.option._
+import scalaz.syntax.traverse._
+import scalaz.{-\/, EitherT, NonEmptyList, \/, \/-}
+import spray.json._
+
+import scala.concurrent.duration.FiniteDuration
+import scala.concurrent.{ExecutionContext, Future}
+import com.daml.logging.{ContextualizedLogger, LoggingContextOf}
+import com.daml.metrics.{Metrics, Timed}
+import akka.http.scaladsl.server.Directives._
+import com.daml.http.endpoints.MeteringReportEndpoint
+import com.daml.ledger.client.services.admin.UserManagementClient
+import com.daml.ledger.client.services.identity.LedgerIdentityClient
+
+import scala.util.control.NonFatal
+
+private[http] final class Parties(
+    routeSetup: RouteSetup,
+    partiesService: PartiesService,
+)(implicit ec: ExecutionContext) {
+  import Parties._
+  import routeSetup._, RouteSetup._
+  import json.JsonProtocol._
+  import util.ErrorOps._
+
+  def allParties(req: HttpRequest)(implicit
+      lc: LoggingContextOf[InstanceUUID with RequestID]
+  ): ET[domain.SyncResponse[List[domain.PartyDetails]]] =
+    proxyWithoutCommand((jwt, _) => partiesService.allParties(jwt))(req)
+      .flatMap(pd => either(pd map (domain.OkResponse(_))))
+
+  def parties(req: HttpRequest)(implicit
+      lc: LoggingContextOf[InstanceUUID with RequestID]
+  ): ET[domain.SyncResponse[List[domain.PartyDetails]]] =
+    proxyWithCommand[NonEmptyList[domain.Party], (Set[domain.PartyDetails], Set[domain.Party])](
+      (jwt, cmd) => partiesService.parties(jwt, toNonEmptySet(cmd))
+    )(req)
+      .map(ps => partiesResponse(parties = ps._1.toList, unknownParties = ps._2.toList))
+
+  def allocateParty(req: HttpRequest)(implicit
+      lc: LoggingContextOf[InstanceUUID with RequestID],
+      metrics: Metrics,
+  ): ET[domain.SyncResponse[domain.PartyDetails]] =
+    EitherT
+      .pure(metrics.daml.HttpJsonApi.allocatePartyThroughput.mark())
+      .flatMap(_ => proxyWithCommand(partiesService.allocate)(req).map(domain.OkResponse(_)))
+}
+
+private[endpoints] object Parties {
+  private def partiesResponse(
+      parties: List[domain.PartyDetails],
+      unknownParties: List[domain.Party],
+  ): domain.SyncResponse[List[domain.PartyDetails]] = {
+
+    val warnings: Option[domain.UnknownParties] =
+      if (unknownParties.isEmpty) None
+      else Some(domain.UnknownParties(unknownParties))
+
+    domain.OkResponse(parties, warnings)
+  }
+}
